+++ Ranger/UI/TourneyTable/TourneyTableView/Cells/Finishes/FinishesGraphCellView.swift
@@ -8,6 +8,7 @@
 
 import Foundation
 import SwiftUI
+import SharkScope
 
 
 class FinishesGraphCellView: PlayerCellView