+++ SharkScope/PlayerSummary/PlayerSummaryRequest.swift
@@ -13,20 +13,21 @@ import Foundation
 /// Requests player summary information. The response contains basic
 /// information about the user as well as all free statistical
 /// information and a small number of most recent tournament results.
-struct PlayerSummaryRequest: Request
+public struct PlayerSummaryRequest: Request
 {
     
     
-    typealias RootResponseType = PlayerSummary
+    public typealias RootResponseType = PlayerSummary
     
     
     let network: String
     let player: String
-    var path: String { "networks/\(network)/players/\(player)" }
-    var parameters: KeyValuePairs<String, String>
+    
+    public var path: String { "networks/\(network)/players/\(player)" }
+    public var parameters: KeyValuePairs<String, String>
     {
         // TODO: Persist filter settings per player.
         (player == "Borbas.Geri") ? [ "filter" : "Date:1Y" ] : [:]
     }
-    var useCache: Bool = true
+    public var useCache: Bool = true
 }