+++ SharkScope/SharkScope.h
@@ -1,0 +1,19 @@
+//
+//  SharkScope.h
+//  SharkScope
+//
+//  Created by Geri Borbás on 2020. 01. 17..
+//  Copyright © 2020. Geri Borbás. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for SharkScope.
+FOUNDATION_EXPORT double SharkScopeVersionNumber;
+
+//! Project version string for SharkScope.
+FOUNDATION_EXPORT const unsigned char SharkScopeVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <SharkScope/PublicHeader.h>
+
+